+++ src/test/groovy/mock/CustomWARPackager.groovy
@@ -1,0 +1,9 @@
+package mock
+
+
+/**
+ * Mock customWARPackager step
+ */
+class CustomWARPackager implements Serializable {
+  public void build(String path, String war, String bom) { }
+}
